backend_udguard/requirements.txt
asgiref==3.8.1
Django==5.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
signals==0.0.2
sqlparse==0.5.3
tzdata==2025.2
